Refactor: new `render` package
