+++ tests/test_view.py
@@ -2,7 +2,7 @@
 
 from pygame import Vector2
 
-from two_d_game_ai.view import View
+from two_d_game_ai.render.view import View
 from two_d_game_ai.world import World
