+++ two_d_game_ai/render/__init__.py
@@ -1,0 +1,1 @@
+"""Package for rendering the model using Pygame."""
